refactor: Minimal LoopMode

+++ rustysynth_test/src/instrument_util.rs
@@ -63,7 +63,7 @@ pub fn check(region: &InstrumentRegion, values: &[f64; 50]) {
     assert!(are_equal(region.get_initial_attenuation() as f64, values[43]));
     assert!(are_equal(region.get_coarse_tune() as f64, values[44]));
     assert!(are_equal(region.get_fine_tune() as f64, values[45]));
-    assert!(are_equal(i16::from(region.get_sample_modes()) as f64, values[46]));
+    assert!(are_equal(region.get_sample_modes() as usize as f64, values[46]));
     assert!(are_equal(region.get_scale_tuning() as f64, values[47]));
     assert!(are_equal(region.get_exclusive_class() as f64, values[48]));
     assert!(are_equal(region.get_root_key() as f64, values[49]));
